docs: moved provider functionality doc snippets
