+++ apps/docs-snippets2/src/provider/functionality/get-blocks.ts
@@ -1,7 +1,7 @@
 // #region getBlocks
 import { Provider } from 'fuels';
 
-import { LOCAL_NETWORK_URL } from '../env';
+import { LOCAL_NETWORK_URL } from '../../env';
 
 const provider = await Provider.create(LOCAL_NETWORK_URL);
