+++ apps/docs-snippets2/src/provider/functionality/get-message-proof.ts
