+++ apps/docs-snippets2/src/provider/functionality/get-resources-to-spend.ts
@@ -2,7 +2,7 @@
 import type { CoinQuantityLike, ExcludeResourcesOption } from 'fuels';
 import { Provider, ScriptTransactionRequest, Wallet } from 'fuels';
 
-import { LOCAL_NETWORK_URL, WALLET_PVT_KEY } from '../env';
+import { LOCAL_NETWORK_URL, WALLET_PVT_KEY } from '../../env';
 
 const provider = await Provider.create(LOCAL_NETWORK_URL);
 const wallet = Wallet.fromPrivateKey(WALLET_PVT_KEY, provider);
